# build/docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
		"examples": {},
		"headers": {},
		"parameters": {},
		"requestBodies": {},
		"responses": {},
		"schemas": {
			"ProductType": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"category": {
						"type": "string"
					},
					"stock": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"id",
					"name",
					"price",
					"category",
					"stock"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProductResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/ProductType"
					}
				},
				"required": [
					"message",
					"data"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProductCreateRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"category": {
						"type": "string"
					},
					"stock": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"price",
					"category",
					"stock"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProductUpdateRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"price": {
						"type": "number",
						"format": "double"
					},
					"category": {
						"type": "string"
					},
					"stock": {
						"type": "number",
						"format": "double"
					}
				},
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "tsoa-express-api",
		"version": "1.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {}
	},
	"paths": {
		"/v1/products": {
			"get": {
				"operationId": "GetAllProducts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"data": {},
										"message": {
											"type": "string"
										}
									},
									"required": [
										"data",
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error"
					}
				},
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateItem",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProductResponse"
								}
							}
						}
					},
					"201": {
						"description": "Created Success"
					},
					"400": {
						"description": "Bad Request"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ProductCreateRequest"
							}
						}
					}
				}
			}
		},
		"/v1/products/{id}": {
			"get": {
				"operationId": "GetItemById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProductResponse"
								}
							}
						}
					},
					"404": {
						"description": "Item not found"
					},
					"500": {
						"description": "Error fetching item"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateItemById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProductResponse"
								}
							}
						}
					},
					"400": {
						"description": "Invalid input"
					},
					"404": {
						"description": "Item not found"
					},
					"500": {
						"description": "Error updating item"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ProductUpdateRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteItemById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Item not found"
					},
					"500": {
						"description": "Error deleting item"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}